Moving code relevant for controller based artifacts into the ControllerBased namespace
